[ACTIVEMODEL] Added support for effective importing data from ActiveModel class into the index, via bulk API
Example:

    class Book < ActiveRecord::Base
      include Slingshot::Model::Search
      include Slingshot::Model::Callbacks
    end

    Book.import

Yes, that's all :)

Your model must provide a `paginate` method (eg. via will_paginate), since we don't want to fry memory with large data.

Any parameters you pass to the `import` method are passed to the `paginate` method.

This is especially handy when you want to limit the documents, eg.:

    Post.paginate(:conditions => {:title => "Search"}, :page => 1, :per_page => 20)

Also, it's very convenient for the excentrical approach of CouchRest Model:

    Post.import(:design_doc => "Post", :view_name => "all", :page => 1, :per_page => 1000, :include_docs => true)