[ACTIVEMODEL] Fixed handling of fields returned prefixed by _source from ES [#31]
The underlying issue is this, reported by @vhyza: <https://github.com/karmi/tire/pull/31#issuecomment-1340967>

For a query with limited fields, such as:

<http://localhost:9200/fields_test/_search?q=shay&fields=message,_source.person>

ES returns JSON like this:

    fields: {
      message: "This is a tweet!"
      _source.person: {
        sid: "12345"
        name: {
          last_name: "Banon"
          first_name: "Shay"
        }
      }
    }

Notice the `_source` prefix for the `person`.
It gets even more tricky with nested hashes such as `_source.track.info.duration`.
This should be converted into a regular Hash and accessible as `track.info.duration.minutes` property of the result.

See the test suite for example data. This commit closes issue #31.