Add support for Boolean Filtered Queries.
 * http://www.elasticsearch.org/guide/reference/query-dsl/bool-filter.html

This patch only implements the `must` logic, and has no facility for the
caching behaviour offered by boolean filtered queries.

This is slightly different to the `filter(:and, { {filter1},
{filter2}})` option introduced in `d2feb840` by @karmi in that it has
different semantics, and doesn't require one to collect a list of
filters (for example in conditional blocks) before calling
`filter(:and,...)`

This patch is incomplete, I will implement `should` and `should_not` as
soon as time becomes available. I don't expect it to be pulled upstream
until those are implemented, but perhaps there is enough value in my
contribution that it will be pulled.