Fix Tire.search with a payload rather than a block.
Previously calling Tire.search with a string payload would cause an
exception to be raised in Tire::Results::Collection, since the payload was
passed directly into the Collection.new options parameter.

Passing any options to Tire.search would result in those options being
encoded directly in the body of the search, causing the search to fail.

This also resulted in missing functionality for Tire.search with a payload,
since options couldn't be specified, including the document type. This
commit adds support for this through specifying the search body in a
payload options, then recognizing additional options supported by
Tire.search with a block.