[ACTIVEMODEL] Fixed and cleaned up URL-escaping of document type
This commit builds on work done in 5151ff9.

It properly URL-encodes document types when storing/removing/retrieving documents in the index
and when executing searches.

It decodes the document types when building the results (`Results::Collection`).

It adds a number of unit and integration tests to check and document the behaviour.

It closes #201, closes #204, closes #264 and closes #271.