Added support for reindexing the index documents into new index
Added a `Tire::Index#reindex` method which uses the "scan/scroll" API to retrieve
documents from index A and import them into index B.

If the new index does not exists, it is created: you may pass settings as an options hash.

You may pass a Tire DSL query as a block to the method.

Example:
--------

    require 'tire'

    Tire.configure { logger STDERR }

    Tire.index('scantest-new') { delete }

    documents = (1..100).map do |i|
      { id: i, type: 'test', title: "Document #{i}" }
    end

    puts "", "Importing data into ElasticSearch..."

    Tire.index 'scantest' do
      delete and create :settings => { :number_of_shards => 1, :number_of_replicas => 0 }
      import documents
      refresh
    end

    Tire.index('scantest').reindex 'scantest-new', settings: { number_of_shards: 3 } do
      query { string '10*' }
    end