Add ability to pass arguments to query { all }
In elasticsearch, match_all takes arguments that affect the boosting.
For example, if you want to perform empty queries that still respect
document boost, you need to pass {'norms_field' => '[field]'} to match_all.

This adds an optional hash argument to Search::Query#all that is a
pass-through to elasticsearch' match_all.