Exit with an error in the Rake import task when the index as not created in elasticsearch
Previously, when the index could not be created due to eg. error in mapping,
it was silently ignored and data were still indexed (with dynamic mapping).

This was a source of hard-to-pin-down errors. When there's an error while
creating the index, the Rake task will exit(1) and print the HTTP response, now.

Closes #392