Added a `match` query type
The `match` query is the most convenient way of exposing a search feature to end-users.

It automatically matches field analyzer with query analyzer, builds a corresponding query,
allows for "as you type" search with the `phrase_prefix` option, etc.

Tire exposes a single DSL method to use the query:

    Tire.search 'people' do
      query do
        match :last_name, 'Smith'
      end
    end

To search in multiple fields (with the `multi_match` query), just use a list of fields as the first argument:

    Tire.search 'people' do
      query do
        match [:first_name, :last_name], 'John'
      end
    end

You may want to give different fields specific boost with the caret notation:

    Tire.search 'people' do
      query do
        match ['first_name', 'last_name^10'], 'John'
      end
    end

The `match` method can be invoked repeatedly, in which case Tire will build a `must` boolean query automatically,
allowing you to quickly create "compound" queries working across multiple field types.

So, the following query:

    Tire.search 'people' do
      query do
        match :last_name, 'John'
        match :age, 30
        match :gender, 'male'
      end
    end

is equivalent to this explicit boolean query:

    Tire.search 'people' do
      query do
        boolean do
          must { match :last_name, 'John' }
          must { match :age, 30           }
          must { match :gender, 'male'    }
        end
      end
    end

See:

* http://www.elasticsearch.org/guide/reference/query-dsl/match-query.html
* http://www.elasticsearch.org/guide/reference/query-dsl/multi-match-query.html