[#525] Fixed bug where the `matches` Tire method interfered with a model method of the same name
In the current [implementation](https://github.com/karmi/tire/blob/v0.5.1/lib/tire/model/search.rb#L196-L202) of `matches`,
the `@attributes` variable of the model has been accessed directly.

This lead to an application crash when new model with a custom implementation of a `matches` method or a `has_many :matches`
declaration was saved:

    NoMethodError: undefined method `each' for nil:NilClass

This patch fixes the bug and preserves the original behaviour. Access the `matches` (percolator results) through the Tire proxy:

    @mymodel.tire.matches

Closes #525, fixes #501, fixes #280.