Added `Tire::Results::Collection#each_with_hit`
Previously, there was no convenient way to access the data returned by Elasticsearch
such as `_score` when the `:load` option was used.

This commit implements the `each_with_hit` method for results collection:

    results = Article.search 'One', :load => true

    results.each_with_hit do |result, hit|
      puts "#{result.title} (score: #{hit['_score']})"
    end

    => One (score: 0.300123)

The model instance is returned in the `result` variable,
the raw Hash from Elasticsearch in the `hit` variable.

Closes #562, closes #406.