Remove the `to_a` in `paginate` method injected in import Rake task
When using Kaminari, whatever is returned from the scopes will behave as an array.

In the case of data_mapper this is especially important, because what gets returned also handles eager loading records. With the to_a in place, it's very likely that an n+1 scenario will be created during import.

Closes #653