[FIX] Fix `Item#to_hash` functionality to work with Arrays
I have extended the nested to_hash method by the functionality to cover
arrays containing to_hash'able elements as well. I could not find a way
to integrate it into the existing statement. Instead, I have added an
additional case specifically for arrays.

I have added a failing unit test case to reveal a problem with the
to_hash method. Previously, the method did not cover the case where
a Results::Item contains an array value with Results::Item objects.
In this case the Results::Item objects were not converted to a Hash.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>

Closes #722