Add test case for an Item with array of hashes
I have added a failing unit test case to reveal a problem with the
to_hash method. Currently, the method does not cover the case where
a Results::Item contains an array value with Results::Item objects.
In this case the Results::Item objects are not converted to a Hash.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>