Allow query to take longer than 10 min.
Snip from the mongodb doc:

By default, the server will automatically close the cursor after 10 minutes of inactivity or if client has exhausted the cursor. To override this behavior, you can specify the noTimeout wire protocol flag in your query; however, you should either close the cursor manually or exhaust the cursor.

http://docs.mongodb.org/manual/core/read-operations/

This especially happens if a lot of data is going to be index.